Use DataSource interface in up command
The cofidectl up command previously used a LocalDataSource, rather than
the DataSource interface. This will become a problem when we try to add
other data sources, i.e. the Connect plugin.

This change fixes the issue by changing to use the DataSource interface.
It requires providing the HelmValuesGenerator with the DataSource
interface to allow it to query TrustZones.

Fixes: #68