Update trust zone with bundle endpoint and bundle
During cofidectl up, we query the bundle endpoint URL and bundle for
each trust zone, in order to allow other trust zones to federate with
them. Previously this information was not being persisted in the
configuration file.

This only worked because of issue #79, since GetTrustZone method returns
a pointer to the TrustZone, so updating the TrustZone's bundle endpoint
and bundle would also update the trust zone in the (in-memory) config.
In order to fix #79 we need to first explicitly update the trust zone in
the config with the bundle endpoint and bundle fields set.

Fixes: #69