Fix waiting for SPIRE server pod & service
The list watches on the pod & service were not previously using a
selector, so returned all pods and services in the spire namespace. This
could lead the wait to end before both the SPIRE server pod & service
are ready.

A second issue is that the isServiceReady function checked for the
presence of a clusterIP, but actually we need the externalIP to be
present.

This change fixes the first issue by applying a field selector that
matches the name of the pod or service. The second issue is fixed by
waiting for the service's external IP to be populated.

Fixes: #66