Add FederatesWith to SPIRE controller manager clusterSPIFFEID entries
In order for workloads to receive trust bundles for federated trust
domains, their clusterSPIFFEID entries in SPIRE controller manager must
include the trust domain in the federatesWith parameter.

This is subtley different than the cofide-cli implementation, in that it
references trust zones rather than trust domains in the --federates-with
argument.

Fixes: #62