Make copies of Config objects across LocalDataSource interface
We should not return pointers to objects in the main Config state of the
LocalDataSource, since that would allow modification of the Config from
outside of the LocalDataSource. We should also not store pointers to
objects provided by a caller of the DataSource interface in the Config,
for the same reason.

This change fixes the issue by implementing clone functions for each of
the main data types that use the proto.Clone function.

Fixes: #79