Remove nested list objects from cofide.yaml
Currently the Config object stores trust zones in a
trust_zone_proto.TrustZoneList and attestation policies in a
attestation_policy_proto.AttestationPolicyList. This results in a nested
structure in cofide.yaml:

  trustzones:
    trustzones:
      - name: tz1 ...
  attestationpolicies:
    policies:
      - name: ap1 ...

It also requires traversing the list objects in code:

  for _, tz := range config.TrustZones.TrustZones

This change fixes the issue by directly using slices of the TrustZone
and AttestationPolicy messages in Config.

Fixes: #86