Store full bundle endpoint URL in config
This ensures that we don't need to duplicate the URL formatting function.

It also highlighted that on the first Helm install we were previously
adding clusterFederatedTrustDomains entries with invalid bundle
endpoints because the BundleEndpointUrl was empty. It started to cause
problems after changing to use the full URL because the bundle endpoint
URL written to Helm values is now empty. To resolve this we avoid adding
clusterFederatedTrustDomains entries until the bundle endpoint URL is
known.