Change federation trust zone names to from and to
A federation was previously composed of left and right trust zones. This
made sense when the relationship was bidirectional, but it is now
unidirectional. The left and right identifiers do not make this obvious.

This change switches the names to "from" and "to".

Depends on: https://github.com/cofide/cofide-api-sdk/pull/15

Fixes: #103