Fall back to SPDY for exec if WebSockets aren't supported
If using cofidectl with an older (pre 1.31.0) k8s cluster, it may fail
to exec into the spire-server container. This is because websockets are
behind a default false feature gate until 1.31.0, and unsupported before
1.29.0.

This change fixes the issue using a fallback executor, adapted from the
exec code in kubectl. It will try to use websockets, then fall back to
SPDY if unavailable.

Fixes: #120