Support per-plugin configuration
For some plugins it would be useful to be able to store configuration
options in cofide.yaml. This change updates the plugin manager to
support setting and retrieving per-plugin configuration keyed by the
plugin name. The config can be set during Init or later using
GetPluginConfig.

A Struct protobuf message is used to allow for arbitrarily structured
plugin config. Plugins can use the Struct directly or convert it to/from
a more convenient native Go type.

Fixes: #54

Depends-On: https://github.com/cofide/cofide-api-sdk/pull/8