Add workload status command
This change adds a 'cofidectl workload status' command. The command
accepts a trust zone, pod name and namespace, and deploys an ephemeral
debug container to the pod. This container emits diagnostic information
about the SVIDs provided to the workload.

This requires us to set disableContainerSelectors=true in the SPIRE Helm
configuration, to allow the debug container to obtain an ID.

Fixes: #14
Co-Authored-By: Matt Bates <matt@cofide.io>
Co-Authored-By: Maartje Eyskens <maartje@cofide.io>