Add --log-level argument to root command
This allows setting the log level threshold. The default is ERROR.
Currently we do not do much logging in cofidectl, but this may change
now that we can set the level.

The main benefit for now is that gRPC plugin log level is also set based
on this argument, giving more insight into plugin issues.

Fixes: #82