Separate Cluster message from TrustZone
This change modifies cofidectl to adapt to the changes in
https://github.com/cofide/cofide-api-sdk/pull/25, in which the TrustZone
message has been refactored to extract per-Kubernetes cluster details
into a Cluster message.

For now we restrict the number of clusters per trust zone to one.
Clusters are not yet exposed by the CLI as separate from trust zones.
They are currently embedded into trust zones in the data model but will
be referenced by name or ID in future.

Fixes: #126