Make Clusters a top-level entity in the data model
This change modifies cofidectl to adapt to the changes in
https://github.com/cofide/cofide-api-sdk/pull/29 in which:

- the Clusters field has been removed from the TrustZone message
- a Clusters field has been added to the Config message
- AddCluster, GetCluster and ListClusters RPCs have been added to the
  DataSourcePluginService

Clusters are now a top-level entity in the data model, rather than
embedded in trust zones. As such, Cluster objects are now retrieved via
the DataSource interface.

The limitation of one cluster per trust zone remains, and clusters are
not exposed in the CLI.

Fixes: #127