feat: support limiting up/down to specific trust zones
This change adds a repeated --trust-zone argument to the cofidectl up
and down commands. When specified, it allows to limit the deploy or tear
down to one or more trust zones. When omitted, the previous behaviour of
deploying or tearing down all trust zones is maintained.

Also adds a --quiet argument to the cofidectl down command, in line with
the up command.

Fixes: #193